Re-work .js files.
